+++ mini_pupper_controllers/mini_pupper_controllers_plugin.xml
@@ -1,0 +1,9 @@
+<library path="simple_quadruped_controller">
+  <class name="mini_pupper_controllers/SimpleQuadrupedController"
+         type="mini_pupper_controllers::SimpleQuadrupedController"
+         base_class_type="controller_interface::ControllerInterface">
+    <description>
+      A minimal controller that holds joints at configured default positions and accepts position commands.
+    </description>
+  </class>
+</library>
